Truncate (double)bwlimit.bucket to (size_t)len

+++ lib/network/tsnetwork_bwlimit.c
@@ -199,9 +199,9 @@ network_bwlimit_get(int op, size_t * len)
 
 	/* Return the appropriate value. */
 	if (op == NETWORK_OP_READ)
-		*len = limit_read.bucket;
+		*len = (size_t)limit_read.bucket;
 	else
-		*len = limit_write.bucket;
+		*len = (size_t)limit_write.bucket;
 
 	/*
 	 * If the allowed bandwidth is less than one normal-sized TCP segment,
